Move new logic to useAccountTotalFiatBalance
