+++ ui/hooks/useAccountTotalFiatBalance.js
@@ -1,9 +1,10 @@
-import { useSelector } from 'react-redux';
+import { shallowEqual, useSelector } from 'react-redux';
 import {
   getAllTokens,
   getCurrentChainId,
   getCurrentCurrency,
   getMetaMaskCachedBalances,
+  getTokenExchangeRates,
 } from '../selectors';
 import {
   getValueFromWeiHex,
@@ -11,6 +12,8 @@ import {
 } from '../../shared/modules/conversion.utils';
 import { getConversionRate } from '../ducks/metamask/metamask';
 import { formatCurrency } from '../helpers/utils/confirm-tx.util';
+import { getTokenFiatAmount } from '../helpers/utils/token-util';
+import { isEqualCaseInsensitive } from '../../shared/modules/string-utils';
 import { useTokenTracker } from './useTokenTracker';
 
 export const useAccountTotalFiatBalance = (
@@ -20,6 +23,11 @@ export const useAccountTotalFiatBalance = (
   const currentChainId = useSelector(getCurrentChainId);
   const conversionRate = useSelector(getConversionRate);
   const currentCurrency = useSelector(getCurrentCurrency);
+
+  const contractExchangeRates = useSelector(
+    getTokenExchangeRates,
+    shallowEqual,
+  );
 
   const cachedBalances = useSelector(getMetaMaskCachedBalances);
   const balance = cachedBalances?.[address] ?? 0;
@@ -40,13 +48,33 @@ export const useAccountTotalFiatBalance = (
     hideZeroBalanceTokens: shouldHideZeroBalanceTokens,
   });
 
+  // Create fiat values for token balances
+  const tokenFiatBalances = tokensWithBalances.map((token) => {
+    const contractExchangeTokenKey = Object.keys(contractExchangeRates).find(
+      (key) => isEqualCaseInsensitive(key, token.address),
+    );
+    const tokenExchangeRate =
+      (contractExchangeTokenKey &&
+        contractExchangeRates[contractExchangeTokenKey]) ??
+      0;
+
+    const totalFiatValue = getTokenFiatAmount(
+      tokenExchangeRate,
+      conversionRate,
+      currentCurrency,
+      token.string,
+      token.symbol,
+      false,
+      false,
+    );
+
+    return totalFiatValue;
+  });
+
   // Total native and token fiat balance as a string (ex: "8.90")
-  const totalTokenFiatBalance = tokensWithBalances.map(
-    (token) => token.totalFiatValue ?? '0',
-  );
   const totalFiatBalance = sumDecimals(
     nativeFiat,
-    ...totalTokenFiatBalance,
+    ...tokenFiatBalances,
   ).toString(10);
 
   // Fiat balance formatted in user's desired currency (ex: "$8.90")
